feat(DTFS2-8310): split up apim gift payload types
