+++ trade-finance-manager-api/server/v1/mappings/apim-gift-payloads/types/apim-gift.ts
@@ -1,42 +1,7 @@
-import { APIM_GIFT_INTEGRATION, COUNTERPARTY_ROLE_CODE, PRODUCT_TYPE_CODES, REPAYMENT_TYPE } from './constants';
-
-export type PartyUrns = {
-  bondGiver?: string;
-  bondBeneficiary?: string;
-  issuingBank?: string;
-};
+import { APIM_GIFT_INTEGRATION, COUNTERPARTY_ROLE_CODE, PRODUCT_TYPE_CODES, REPAYMENT_TYPE } from '../constants';
 
 type CreditTypeMap = (typeof APIM_GIFT_INTEGRATION)['DEFAULTS']['OVERVIEW']['CREDIT_TYPE'];
 type CreditTypeValue = CreditTypeMap[keyof CreditTypeMap];
-
-export type AmendmentAmountDataParams = {
-  amount: number;
-  date: string;
-};
-
-export type AmendmentDataPayloadIncreaseOrDecreaseAmount = {
-  amount: number;
-  date: string;
-};
-
-export type AmendmentPayloadDecreaseAmount = {
-  amendmentType: typeof APIM_GIFT_INTEGRATION.AMENDMENT_TYPE.DECREASE_AMOUNT;
-  amendmentData: AmendmentDataPayloadIncreaseOrDecreaseAmount;
-};
-
-export type AmendmentPayloadIncreaseAmount = {
-  amendmentType: typeof APIM_GIFT_INTEGRATION.AMENDMENT_TYPE.INCREASE_AMOUNT;
-  amendmentData: AmendmentDataPayloadIncreaseOrDecreaseAmount;
-};
-
-export type AmendmentDataPayloadReplaceExpiryDate = {
-  expiryDate: string;
-};
-
-export type AmendmentPayloadReplaceExpiryDate = {
-  amendmentType: typeof APIM_GIFT_INTEGRATION.AMENDMENT_TYPE.REPLACE_EXPIRY_DATE;
-  amendmentData: AmendmentDataPayloadReplaceExpiryDate;
-};
 
 export type ApimGiftProductTypeCode = (typeof PRODUCT_TYPE_CODES)[keyof typeof PRODUCT_TYPE_CODES];
