+++ trade-finance-manager-api/server/v1/mappings/apim-gift-payloads/types/party-urns.ts
@@ -1,0 +1,5 @@
+export type PartyUrns = {
+  bondGiver?: string;
+  bondBeneficiary?: string;
+  issuingBank?: string;
+};
